renamed to syftcli
