+++ packages/syftcli/.gitignore
