+++ packages/syftcli/scripts/build-cli.sh
