+++ packages/syftcli/src/cli.py
@@ -2,7 +2,7 @@
 from rich import box
 from rich.console import Console
 from rich.table import Table
-from syft_cli.version import __version__
+from src.version import __version__
 import typer
 
 app = typer.Typer(name="Syft CLI", no_args_is_help=True)
