+++ packages/syftcli/tests/version_test.py
